Switch to Pipeline.validate_config/1 for consistency
There was no need to validate and parse in the same step, and by
renaming this, it reflects the same function as implemented by the
providers as well. Makes everything a bit cleaner.

Note, dialyzer seems happy for the moment, but I theorize that once
we're actually creating pipelines, the `:ok | no_return` might cause
some issues that will have to be cleaned up.