Add an initial MergeRequest module with FSM
To keep things simple, I'm trying to just use the struct and pattern matching
in order to implement the state changes. This may have to be tweaked later on,
but seems like a straight-forward way to go without having to couple state
transitions to cross-process communication via `:gen_fsm` or `:gen_statem`.

References:
* https://github.com/sasa1977/fsm/issues/3
* [The missing finite state machine, Handling a finite state machine with isolated side effects and without erlangs gen_fsm](https://groups.google.com/d/topic/elixir-lang-talk/Ac3B1Dj8br0/discussion)