Add the initial /webhook listener process
There's a lot going on in this commit, but this is still just the
framework to start handling incoming HTTP callbacks...

First, we're setting defaults for the required environment configuration
settings, and then allowing for overrides in the *config/* hierarchy.
I'll be updating the rest of the code to only raise exceptions for
settings that are truly required and are missing/invalid. I've left
a commented-out version of the config syntax in *config.exs* as an
example of how to override.

Second, we're using a feature flag `enable_http_workers` to disable the
listeners when running in the `:test` environment. I believe we'll be
able to test the Plug code without them, and it prevents port conflict
issues if the app was already up and running. I also need to be
cognizant of workers that will be hitting external services in relation
to how the tests run.

Last, the Plug.Router currently doesn't do anything beyond send the
desired responses for connections. By halting after the default match,
I'm hoping to prevent the JSON parser from running unnecessarily.

References:
* [Not starting workers in test environment.](https://groups.google.com/d/topic/elixir-lang-talk/T5a4vdMM1U8/discussion)