Pass valid JSON webhook events to a handler module
This still technically parses the body for all requests, as I couldn't
come up with a clean way to only run the JSON decoder on the "/webhook"
match. But, with the addition of `Plug.ErrorHandler`, the failure mode
for bad requests is much nicer, and we still send a 404 Not Found if
things get to that point.