Use GenStage for dispatching webhook events
This is an initial implementation that uses a broadcaster that will send all
webhook events to any subscribers. The subscribers in this case will be the
providers. We'll see how this shakes out, but the functional replacement of the
previous handler stub code is there.

Most of the code was copied/pasted from the GenStage announcement.

References:
* http://elixir-lang.org/blog/2016/07/14/announcing-genstage
* http://www.elixirfbp.org/2016/07/genstage-example.html
* https://hexdocs.pm/gen_stage/Experimental.GenStage.html