Remove the queue module and store MRs as a map in the pipeline
To make things more simple to start, I think I'll try relying on the approval
time of the merge requests as an ad-hoc queue rather than try to maintain state
in two separate data structures. If this becomes a problem, I'll address that
later, but I needed a map of the MRs anyway to make updates to the state in an
efficient manner (as you couldn't easily update the state of items in the
middle of the queue).