Implement handling of comments
This also includes a large reorganization. I've tried to simplify a lot
of variable names to "mreqs" instead of the verbose "merge_requests".
Some are still left for clarity.

I've also moved the command parsing into the pipeline module (which
keeps getting bigger), as it didn't really make sense in the Comment
module, as it's just processing any generic text and might be useful for
a CLI interface or something down the road. I'll think about how to
split the pipeline module up once it gets unmanageable.