Reformat mix.exs files to a more readable style
I've never been a fan of having the enclosing brackets for lists on the
same lines as items. I also like to have the trailing comma on every
item to make maintenance easier [1]. Always including parentheses for
local zero-arity function calls and definitions [2] is also something
I've been trying to get into the habit of doing.

[1] https://github.com/lexmag/elixir-style-guide#trailing-comma
[2] https://github.com/lexmag/elixir-style-guide#zero-arity-parens