Use the verb "get" for consistency
Rather that use the inconsistent verbs that are in place in the upstream
API documentation, I think it makes sense to standardize on using "get"
for pure GET requests, which also is familiar to other usages of the
term throughout Elixir as a language. The verbs "fetch" and "show"
either have different meanings in Elixir, or are not used, so I think
this will make the wrapper feel more native.