WIP refactoring of how the API should work
I'm going to simplify and only have the singular-named modules rather
than singular for the struct and plural for the functions. I'll control
the free-form input at as low of a place as possible, the make
discovering of other attributes like repository branches and keys
require a %Repo{} struct in order to perform the query.

A lot will be changing here, but I wanted to capture the current state.