Use Registry instead of gproc for process registration
This removes the gproc dependency now that Registry is a part of the
core language. I think I can simplify the pipeline code now that I've
been touching it, but wanted to get a straight conversion working as
a first step.

This does require bumping up the minimum required version for Elixir.